Move autocomplete options to package
